feat: Implement signal parsing.
