+++ crates/gdtk-parser/src/functions.rs
@@ -4,50 +4,17 @@ use gdtk_ast::poor::{ASTFunction, ASTStatement};
 use gdtk_lexer::{Token, TokenKind};
 
 use crate::block::parse_block;
-use crate::utils::{expect_blank_prefixed, peek_non_blank};
+use crate::utils::{collect_params, expect_blank_prefixed, peek_non_blank};
 use crate::values::parse_value;
-use crate::variables::parse_variable;
 
 pub fn parse_func<'a, T>(iter: &mut Peekable<T>) -> ASTStatement<'a>
 where
     T: Iterator<Item = Token<'a>>,
 {
     let identifier = expect_blank_prefixed!(iter, TokenKind::Identifier(s), s);
-    let mut parameters = vec![];
     let mut return_type = None;
 
-    expect_blank_prefixed!(iter, TokenKind::OpeningParenthesis, ());
-
-    if !matches!(peek_non_blank!(iter).kind, TokenKind::ClosingParenthesis) {
-        loop {
-            if !matches!(peek_non_blank!(iter).kind, TokenKind::Identifier(_)) {
-                panic!("unexpected {:?}, expected function parameter", iter.next());
-            }
-
-            let param = parse_variable(iter, gdtk_ast::poor::ASTVariableKind::FunctionParameter);
-            parameters.push(param);
-
-            match peek_non_blank!(iter) {
-                Token {
-                    kind: TokenKind::Comma,
-                    ..
-                } => {
-                    iter.next();
-                    continue;
-                }
-                Token {
-                    kind: TokenKind::ClosingParenthesis,
-                    ..
-                } => {
-                    iter.next();
-                    break;
-                }
-                other => panic!("unexpected {other:?}, expected a comma or a closing parenthesis"),
-            }
-        }
-    } else {
-        iter.next();
-    }
+    let parameters = collect_params(iter);
 
     if matches!(peek_non_blank!(iter).kind, TokenKind::Arrow) {
         iter.next();
