+++ quick.gd
@@ -1,2 +1,8 @@
 func function() -> void:
 	pass
+
+func function(param1: int = 2):
+	pass
+
+signal frame()
+signal name_changed(name: str)
